% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileUtils.R
\name{composeString}
\alias{composeString}
\title{composes a text adding all the strings passed as
argument}
\usage{
composeString(objects, sep = NULL)
}
\arguments{
\item{objects}{data to convert to text}

\item{sep}{separator to use}
}
\value{
string with the information
}
\description{
composes a text adding all the strings passed as
argument
}
