% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileUtils.R
\name{composeStringForList}
\alias{composeStringForList}
\title{composes a text adding all the strings passed as
argument}
\usage{
composeStringForList(collection, sep = NULL)
}
\arguments{
\item{collection}{strings to compose into a single string}

\item{sep}{separator to use}
}
\value{
composed string
}
\description{
composes a text adding all the strings passed as
argument
}
