% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileUtils.R
\name{composeStringForValues}
\alias{composeStringForValues}
\title{composes a text adding all the strings passed as
argument as values with format}
\usage{
composeStringForValues(objects, sep = NULL)
}
\arguments{
\item{objects}{objects to convert to string}

\item{sep}{separator to use}
}
\value{
string with the composed information
}
\description{
composes a text adding all the strings passed as
argument as values with format
}
