% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphUtils.R
\name{createBnlearnString}
\alias{createBnlearnString}
\title{creates a string describing the structure of the net
example ([A|B:C:F][B][C][D][E][F])}
\usage{
createBnlearnString(variables, parents)
}
\arguments{
\item{variables}{lits of variables involved}

\item{parents}{list of parents for each variable}
}
\description{
creates a string describing the structure of the net
example ([A|B:C:F][B][C][D][E][F])
}
