% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphUtils.R
\name{createTestMatrix}
\alias{createTestMatrix}
\title{creates a test matrix for testing purposes}
\usage{
createTestMatrix(names)
}
\arguments{
\item{names}{variable names for printing the
graph}
}
\description{
creates a test matrix for testing purposes
}
