% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileUtils.R
\name{createTraceFolder}
\alias{createTraceFolder}
\title{creates a new folder (if needed) for storing debug
information about the execution with a given file}
\usage{
createTraceFolder(dbname, maxIter, nParentSets)
}
\arguments{
\item{dbname}{name of data to consider}

\item{maxIter}{number of iterations of the experiment to trace}

\item{nParentSets}{number of parent sets of the experiment}
}
\value{
path to the created folder
}
\description{
creates a new folder (if needed) for storing debug
information about the execution with a given file
}
