% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSEvolutionInfo.R
\name{EvolutionInfo}
\alias{EvolutionInfo}
\title{Class for storing debug info about the execution}
\description{
Class for storing debug info about the execution

Class for storing debug info about the execution
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EvolutionInfo-new}{\code{EvolutionInfo$new()}}
\item \href{#method-EvolutionInfo-addStartText}{\code{EvolutionInfo$addStartText()}}
\item \href{#method-EvolutionInfo-addEndText}{\code{EvolutionInfo$addEndText()}}
\item \href{#method-EvolutionInfo-addText}{\code{EvolutionInfo$addText()}}
\item \href{#method-EvolutionInfo-generateBoundsFile}{\code{EvolutionInfo$generateBoundsFile()}}
\item \href{#method-EvolutionInfo-generateBoundsGraphic}{\code{EvolutionInfo$generateBoundsGraphic()}}
\item \href{#method-EvolutionInfo-generateAvgBoundsGraphic}{\code{EvolutionInfo$generateAvgBoundsGraphic()}}
\item \href{#method-EvolutionInfo-generateBayBoundsGraphic}{\code{EvolutionInfo$generateBayBoundsGraphic()}}
\item \href{#method-EvolutionInfo-generateEntropyGraphic}{\code{EvolutionInfo$generateEntropyGraphic()}}
\item \href{#method-EvolutionInfo-getText}{\code{EvolutionInfo$getText()}}
\item \href{#method-EvolutionInfo-storeInitialBounds}{\code{EvolutionInfo$storeInitialBounds()}}
\item \href{#method-EvolutionInfo-getBestBound}{\code{EvolutionInfo$getBestBound()}}
\item \href{#method-EvolutionInfo-getBestAvgBound}{\code{EvolutionInfo$getBestAvgBound()}}
\item \href{#method-EvolutionInfo-getBestBayBound}{\code{EvolutionInfo$getBestBayBound()}}
\item \href{#method-EvolutionInfo-storeScoreForVariable}{\code{EvolutionInfo$storeScoreForVariable()}}
\item \href{#method-EvolutionInfo-storeScoreForVariables}{\code{EvolutionInfo$storeScoreForVariables()}}
\item \href{#method-EvolutionInfo-storeAvgScoreForVariable}{\code{EvolutionInfo$storeAvgScoreForVariable()}}
\item \href{#method-EvolutionInfo-storeAvgScoreForVariables}{\code{EvolutionInfo$storeAvgScoreForVariables()}}
\item \href{#method-EvolutionInfo-storeBayScoreForVariable}{\code{EvolutionInfo$storeBayScoreForVariable()}}
\item \href{#method-EvolutionInfo-storeBayScoreForVariables}{\code{EvolutionInfo$storeBayScoreForVariables()}}
\item \href{#method-EvolutionInfo-storeEvolutionInfo}{\code{EvolutionInfo$storeEvolutionInfo()}}
\item \href{#method-EvolutionInfo-getLastScore}{\code{EvolutionInfo$getLastScore()}}
\item \href{#method-EvolutionInfo-getLastAvgScore}{\code{EvolutionInfo$getLastAvgScore()}}
\item \href{#method-EvolutionInfo-getLastBayScore}{\code{EvolutionInfo$getLastBayScore()}}
\item \href{#method-EvolutionInfo-getLastScoreForVariable}{\code{EvolutionInfo$getLastScoreForVariable()}}
\item \href{#method-EvolutionInfo-getLastAvgScoreForVariable}{\code{EvolutionInfo$getLastAvgScoreForVariable()}}
\item \href{#method-EvolutionInfo-getLastBayScoreForVariable}{\code{EvolutionInfo$getLastBayScoreForVariable()}}
\item \href{#method-EvolutionInfo-getScoresForVariable}{\code{EvolutionInfo$getScoresForVariable()}}
\item \href{#method-EvolutionInfo-getLastEntropy}{\code{EvolutionInfo$getLastEntropy()}}
\item \href{#method-EvolutionInfo-closeTrace}{\code{EvolutionInfo$closeTrace()}}
\item \href{#method-EvolutionInfo-clone}{\code{EvolutionInfo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EvolutionInfo-new"></a>}}
\if{latex}{\out{\hypertarget{method-EvolutionInfo-new}{}}}
\subsection{Method \code{new()}}{
class constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EvolutionInfo$new(
  filename,
  initializationMethod,
  maxIter,
  nParentSets,
  variables
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{name of file to generate}

\item{\code{initializationMethod}}{id of the method used for initialization}

\item{\code{maxIter}}{maximum number of iterations}

\item{\code{nParentSets}}{number of configurations for parent sets}

\item{\code{variables}}{variables of the graph under analysis}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EvolutionInfo-addStartText"></a>}}
\if{latex}{\out{\hypertarget{method-EvolutionInfo-addStartText}{}}}
\subsection{Method \code{addStartText()}}{
adds text with debug info about begining a method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EvolutionInfo$addStartText(methodName, argumentNames, argumentValues)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{methodName}}{name of the method}

\item{\code{argumentNames}}{a vector with argument names}

\item{\code{argumentValues}}{vector with argument values}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EvolutionInfo-addEndText"></a>}}
\if{latex}{\out{\hypertarget{method-EvolutionInfo-addEndText}{}}}
\subsection{Method \code{addEndText()}}{
adds text with debug info about exiting a method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EvolutionInfo$addEndText(methodName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{methodName}}{name of the method just finishing}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EvolutionInfo-addText"></a>}}
\if{latex}{\out{\hypertarget{method-EvolutionInfo-addText}{}}}
\subsection{Method \code{addText()}}{
add regular text to trace
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EvolutionInfo$addText(message)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message}}{text to add to the information to be shown}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EvolutionInfo-generateBoundsFile"></a>}}
\if{latex}{\out{\hypertarget{method-EvolutionInfo-generateBoundsFile}{}}}
\subsection{Method \code{generateBoundsFile()}}{
generates a file with the data of bounds
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EvolutionInfo$generateBoundsFile(
  filename,
  maxIter,
  nParentSets,
  bnlearnTrain,
  bnlearnTest,
  finalTestScore,
  finalAvgScore,
  finalBayScore
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{name of the data set employed}

\item{\code{maxIter}}{iterations of the experiment}

\item{\code{nParentSets}}{number of parent sets employed}

\item{\code{bnlearnTrain}}{score of bnlearn model on train dataset}

\item{\code{bnlearnTest}}{score of bnlearn model on test dataset}

\item{\code{finalTestScore}}{score of our model on test dataset}

\item{\code{finalAvgScore}}{final average score on test}

\item{\code{finalBayScore}}{final bayesian score on test}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EvolutionInfo-generateBoundsGraphic"></a>}}
\if{latex}{\out{\hypertarget{method-EvolutionInfo-generateBoundsGraphic}{}}}
\subsection{Method \code{generateBoundsGraphic()}}{
generates a graphic with the evolution of the bounds obtained
during the search process
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EvolutionInfo$generateBoundsGraphic(
  iteration,
  bnlearnTrain = -Inf,
  bnlearnTest = -Inf,
  finalTestScore = -Inf
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{iteration}}{number of iteration where the graphics is generated}

\item{\code{bnlearnTrain}}{bnlearn score on train data}

\item{\code{bnlearnTest}}{bnlearn score on test data}

\item{\code{finalTestScore}}{final value of bound on test data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EvolutionInfo-generateAvgBoundsGraphic"></a>}}
\if{latex}{\out{\hypertarget{method-EvolutionInfo-generateAvgBoundsGraphic}{}}}
\subsection{Method \code{generateAvgBoundsGraphic()}}{
generates a graphic with the evolution of the averaging bounds obtained
during the search process
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EvolutionInfo$generateAvgBoundsGraphic(
  iteration,
  bnlearnTrain = -Inf,
  bnlearnTest = -Inf,
  finalTestScore = -Inf
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{iteration}}{number of iteration where the graphics is generated}

\item{\code{bnlearnTrain}}{bnlearn score on train data}

\item{\code{bnlearnTest}}{bnlearn score on test data}

\item{\code{finalTestScore}}{final value of averaging on test data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EvolutionInfo-generateBayBoundsGraphic"></a>}}
\if{latex}{\out{\hypertarget{method-EvolutionInfo-generateBayBoundsGraphic}{}}}
\subsection{Method \code{generateBayBoundsGraphic()}}{
generates a graphic with the evolution of the Bayesian bounds obtained
during the search process
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EvolutionInfo$generateBayBoundsGraphic(
  iteration,
  bnlearnTrain = -Inf,
  bnlearnTest = -Inf,
  finalTestScore = -Inf
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{iteration}}{number of iteration where the graphics is generated}

\item{\code{bnlearnTrain}}{bnlearn score on train data}

\item{\code{bnlearnTest}}{bnlearn score on test data}

\item{\code{finalTestScore}}{final value of averaging on test data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EvolutionInfo-generateEntropyGraphic"></a>}}
\if{latex}{\out{\hypertarget{method-EvolutionInfo-generateEntropyGraphic}{}}}
\subsection{Method \code{generateEntropyGraphic()}}{
generates a graphic with the evolution of the bounds obtained
during the search process
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EvolutionInfo$generateEntropyGraphic()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EvolutionInfo-getText"></a>}}
\if{latex}{\out{\hypertarget{method-EvolutionInfo-getText}{}}}
\subsection{Method \code{getText()}}{
return the text stored with debg information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EvolutionInfo$getText()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EvolutionInfo-storeInitialBounds"></a>}}
\if{latex}{\out{\hypertarget{method-EvolutionInfo-storeInitialBounds}{}}}
\subsection{Method \code{storeInitialBounds()}}{
stores initial information about bounds
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EvolutionInfo$storeInitialBounds(bound, averagingBound, bayesianBound)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bound}}{initial value for bound and bestBound}

\item{\code{averagingBound}}{initial value for averagingBound}

\item{\code{bayesianBound}}{initial value for bayesianBound}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EvolutionInfo-getBestBound"></a>}}
\if{latex}{\out{\hypertarget{method-EvolutionInfo-getBestBound}{}}}
\subsection{Method \code{getBestBound()}}{
gets the best bound from the array of best bounds
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EvolutionInfo$getBestBound()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EvolutionInfo-getBestAvgBound"></a>}}
\if{latex}{\out{\hypertarget{method-EvolutionInfo-getBestAvgBound}{}}}
\subsection{Method \code{getBestAvgBound()}}{
gets the best averaging bound from the array of averaging bounds
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EvolutionInfo$getBestAvgBound()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EvolutionInfo-getBestBayBound"></a>}}
\if{latex}{\out{\hypertarget{method-EvolutionInfo-getBestBayBound}{}}}
\subsection{Method \code{getBestBayBound()}}{
gets the best bayesian bound from the array of averaging bounds
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EvolutionInfo$getBestBayBound()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EvolutionInfo-storeScoreForVariable"></a>}}
\if{latex}{\out{\hypertarget{method-EvolutionInfo-storeScoreForVariable}{}}}
\subsection{Method \code{storeScoreForVariable()}}{
stores a new measure of score for a certain variable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EvolutionInfo$storeScoreForVariable(varIndex, score)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{varIndex}}{index of the variable of interest}

\item{\code{score}}{value of the score to store}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EvolutionInfo-storeScoreForVariables"></a>}}
\if{latex}{\out{\hypertarget{method-EvolutionInfo-storeScoreForVariables}{}}}
\subsection{Method \code{storeScoreForVariables()}}{
stores a new measure of score for a pair of variables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EvolutionInfo$storeScoreForVariables(varIndex1, varIndex2, scores)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{varIndex1}}{index of the first variable of interest}

\item{\code{varIndex2}}{index of the second target variable}

\item{\code{scores}}{values of the scores to store}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EvolutionInfo-storeAvgScoreForVariable"></a>}}
\if{latex}{\out{\hypertarget{method-EvolutionInfo-storeAvgScoreForVariable}{}}}
\subsection{Method \code{storeAvgScoreForVariable()}}{
stores a new measure of averaging score for a certain variable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EvolutionInfo$storeAvgScoreForVariable(varIndex, score)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{varIndex}}{index of the variable of interest}

\item{\code{score}}{value of the score to store}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EvolutionInfo-storeAvgScoreForVariables"></a>}}
\if{latex}{\out{\hypertarget{method-EvolutionInfo-storeAvgScoreForVariables}{}}}
\subsection{Method \code{storeAvgScoreForVariables()}}{
stores a new measure of averaging score for a pair of variables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EvolutionInfo$storeAvgScoreForVariables(varIndex1, varIndex2, scores)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{varIndex1}}{index of the first variable of interest}

\item{\code{varIndex2}}{index of the second target variable}

\item{\code{scores}}{values of the scores to store}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EvolutionInfo-storeBayScoreForVariable"></a>}}
\if{latex}{\out{\hypertarget{method-EvolutionInfo-storeBayScoreForVariable}{}}}
\subsection{Method \code{storeBayScoreForVariable()}}{
stores a new measure of bayesian score for a certain variable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EvolutionInfo$storeBayScoreForVariable(varIndex, score)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{varIndex}}{index of the variable of interest}

\item{\code{score}}{value of the score to store}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EvolutionInfo-storeBayScoreForVariables"></a>}}
\if{latex}{\out{\hypertarget{method-EvolutionInfo-storeBayScoreForVariables}{}}}
\subsection{Method \code{storeBayScoreForVariables()}}{
stores a new measure of bayesian score for a pair of variables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EvolutionInfo$storeBayScoreForVariables(varIndex1, varIndex2, scores)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{varIndex1}}{index of the first variable of interest}

\item{\code{varIndex2}}{index of the second target variable}

\item{\code{scores}}{values of the scores to store}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EvolutionInfo-storeEvolutionInfo"></a>}}
\if{latex}{\out{\hypertarget{method-EvolutionInfo-storeEvolutionInfo}{}}}
\subsection{Method \code{storeEvolutionInfo()}}{
stores information about the evolution of the system in order
to analyze its performance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EvolutionInfo$storeEvolutionInfo(entropy, bound, averagingBound, bayesianBound)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entropy}}{entropy value}

\item{\code{bound}}{bound for current iteration}

\item{\code{averagingBound}}{averaging bound for the current iteration}

\item{\code{bayesianBound}}{Bayesian bound for the current iteration}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EvolutionInfo-getLastScore"></a>}}
\if{latex}{\out{\hypertarget{method-EvolutionInfo-getLastScore}{}}}
\subsection{Method \code{getLastScore()}}{
gets the score for the last iteration
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EvolutionInfo$getLastScore()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EvolutionInfo-getLastAvgScore"></a>}}
\if{latex}{\out{\hypertarget{method-EvolutionInfo-getLastAvgScore}{}}}
\subsection{Method \code{getLastAvgScore()}}{
gets the averaging score for the last iteration
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EvolutionInfo$getLastAvgScore()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EvolutionInfo-getLastBayScore"></a>}}
\if{latex}{\out{\hypertarget{method-EvolutionInfo-getLastBayScore}{}}}
\subsection{Method \code{getLastBayScore()}}{
gets the bayesian score for the last iteration
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EvolutionInfo$getLastBayScore()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EvolutionInfo-getLastScoreForVariable"></a>}}
\if{latex}{\out{\hypertarget{method-EvolutionInfo-getLastScoreForVariable}{}}}
\subsection{Method \code{getLastScoreForVariable()}}{
gets the last score for a given variable and a certain
iteration
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EvolutionInfo$getLastScoreForVariable(index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{index of the target variable}

\item{\code{iteration}}{target iteration}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EvolutionInfo-getLastAvgScoreForVariable"></a>}}
\if{latex}{\out{\hypertarget{method-EvolutionInfo-getLastAvgScoreForVariable}{}}}
\subsection{Method \code{getLastAvgScoreForVariable()}}{
gets the last averaging score for a given variable and a certain
iteration
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EvolutionInfo$getLastAvgScoreForVariable(index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{index of the target variable}

\item{\code{iteration}}{target iteration}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EvolutionInfo-getLastBayScoreForVariable"></a>}}
\if{latex}{\out{\hypertarget{method-EvolutionInfo-getLastBayScoreForVariable}{}}}
\subsection{Method \code{getLastBayScoreForVariable()}}{
gets the last bayesian score for a given variable and a certain
iteration
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EvolutionInfo$getLastBayScoreForVariable(index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{index of the target variable}

\item{\code{iteration}}{target iteration}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EvolutionInfo-getScoresForVariable"></a>}}
\if{latex}{\out{\hypertarget{method-EvolutionInfo-getScoresForVariable}{}}}
\subsection{Method \code{getScoresForVariable()}}{
gets the scores for a given variable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EvolutionInfo$getScoresForVariable(index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{index of the target variable}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EvolutionInfo-getLastEntropy"></a>}}
\if{latex}{\out{\hypertarget{method-EvolutionInfo-getLastEntropy}{}}}
\subsection{Method \code{getLastEntropy()}}{
gets the last value of entropy stored
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EvolutionInfo$getLastEntropy()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EvolutionInfo-closeTrace"></a>}}
\if{latex}{\out{\hypertarget{method-EvolutionInfo-closeTrace}{}}}
\subsection{Method \code{closeTrace()}}{
close file
#################################################################
private part of the class
#################################################################
data members description and private methods
for each iteration
bouds obtained for each iteration
bouds obtained for each iteration
folder where trace files will be stored
under execution (entries in stack) in order to
tabulate the text
creates a string describing the structure of the net
creates the file for the text
decrements the counter storing stack level
increments the counter storing stack level
writes the current content of text and clear its content
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EvolutionInfo$closeTrace()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EvolutionInfo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EvolutionInfo-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EvolutionInfo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
