% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileUtils.R
\name{extractsNetName}
\alias{extractsNetName}
\title{return the name of the net from dbname}
\usage{
extractsNetName(dbname)
}
\arguments{
\item{dbname}{name of the data set}
}
\value{
name of the net related to the data file
}
\description{
return the name of the net from dbname
}
