% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileUtils.R
\name{generateDataBases}
\alias{generateDataBases}
\title{generate databases from a net file}
\usage{
generateDataBases(netname, nsamples, nfiles)
}
\arguments{
\item{netname}{name of net}

\item{nsamples}{number of samples to generate}

\item{nfiles}{number of files to generate}
}
\description{
generate databases from a net file
}
