% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphUtils.R
\name{generateGraphFromBnet}
\alias{generateGraphFromBnet}
\title{generates a graph from a net with bnlearn format}
\usage{
generateGraphFromBnet(bnet)
}
\arguments{
\item{bnet}{object of bnet class}
}
\description{
generates a graph from a net with bnlearn format
}
