% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphUtils.R
\name{generateGraphFromMatrix}
\alias{generateGraphFromMatrix}
\title{generates a graph from an adjacency matrix}
\usage{
generateGraphFromMatrix(amat, variables)
}
\arguments{
\item{amat}{adjacency matrix}

\item{variables}{vector of variables}
}
\description{
generates a graph from an adjacency matrix
}
