% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphUtils.R
\name{generatePredominantGraph}
\alias{generatePredominantGraph}
\title{generates the predominant graph reading the highest values from
the matrix of probabilities passed as argument}
\usage{
generatePredominantGraph(probMatrix)
}
\arguments{
\item{probMatrix}{matrix with probabilities}
}
\description{
generates the predominant graph reading the highest values from
the matrix of probabilities passed as argument
}
