% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphUtils.R
\name{generateProbGraph}
\alias{generateProbGraph}
\title{generates a probability graph}
\usage{
generateProbGraph(probMatrix, threshold)
}
\arguments{
\item{probMatrix}{matrix with the probabilistic matrix
generated with the learning algorithm}

\item{threshold}{min value of prob for showing an edge}
}
\description{
generates a probability graph
}
