% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphUtils.R
\name{generateRealNetGraph}
\alias{generateRealNetGraph}
\title{general method for generating a graph with the
real net}
\usage{
generateRealNetGraph(name)
}
\arguments{
\item{name}{name of the bnet to graph}
}
\description{
general method for generating a graph with the
real net
}
