% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileUtils.R
\name{matrixToString}
\alias{matrixToString}
\title{generates a string with the content of a matrix}
\usage{
matrixToString(mat)
}
\arguments{
\item{mat}{matrix to convert into string}
}
\value{
string with matrix information
}
\description{
generates a string with the content of a matrix
}
