% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphUtils.R
\name{plotEvolutionNets}
\alias{plotEvolutionNets}
\title{plot in a single graph the predominant model and the probabilistic
one}
\usage{
plotEvolutionNets(name, probModel, threshold)
}
\arguments{
\item{name}{name of the network of interest}

\item{probModel}{probabilistic matrix generated with
the learning algorithm}

\item{threshold}{probability threshold; only values over
threshold will be printed}
}
\description{
plot in a single graph the predominant model and the probabilistic
one
}
