% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphUtils.R
\name{plotNets}
\alias{plotNets}
\title{plot in a single graph the real model, the predominant
model and the probabilistic and complete model}
\usage{
plotNets(name, bnlearnModel, probModel, threshold, dataPath)
}
\arguments{
\item{name}{name of the network of interest}

\item{bnlearnModel}{model  generated with the initialization
process}

\item{probModel}{probabilistic matrix generated with
the learning algorithm}

\item{threshold}{probability threshold; only values over
threshold will be printed}

\item{dataPath}{path for data folder}
}
\description{
plot in a single graph the real model, the predominant
model and the probabilistic and complete model
}
