% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QMatrix.R
\name{QMatrix}
\alias{QMatrix}
\title{class for storing the values required for learning}
\description{
class for storing the values required for learning

class for storing the values required for learning
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-QMatrix-new}{\code{QMatrix$new()}}
\item \href{#method-QMatrix-copy}{\code{QMatrix$copy()}}
\item \href{#method-QMatrix-getProbFromTo}{\code{QMatrix$getProbFromTo()}}
\item \href{#method-QMatrix-generateParentsList}{\code{QMatrix$generateParentsList()}}
\item \href{#method-QMatrix-getAdjacencyMatrix}{\code{QMatrix$getAdjacencyMatrix()}}
\item \href{#method-QMatrix-getDebug}{\code{QMatrix$getDebug()}}
\item \href{#method-QMatrix-getDimension}{\code{QMatrix$getDimension()}}
\item \href{#method-QMatrix-getIndependence}{\code{QMatrix$getIndependence()}}
\item \href{#method-QMatrix-getState}{\code{QMatrix$getState()}}
\item \href{#method-QMatrix-getValuesFromNames}{\code{QMatrix$getValuesFromNames()}}
\item \href{#method-QMatrix-getValuesFromIndices}{\code{QMatrix$getValuesFromIndices()}}
\item \href{#method-QMatrix-getVariables}{\code{QMatrix$getVariables()}}
\item \href{#method-QMatrix-modelInitialization}{\code{QMatrix$modelInitialization()}}
\item \href{#method-QMatrix-normalizeWithScores}{\code{QMatrix$normalizeWithScores()}}
\item \href{#method-QMatrix-randomInitialization}{\code{QMatrix$randomInitialization()}}
\item \href{#method-QMatrix-save}{\code{QMatrix$save()}}
\item \href{#method-QMatrix-setValue}{\code{QMatrix$setValue()}}
\item \href{#method-QMatrix-show}{\code{QMatrix$show()}}
\item \href{#method-QMatrix-uniformInitialization}{\code{QMatrix$uniformInitialization()}}
\item \href{#method-QMatrix-clone}{\code{QMatrix$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QMatrix-new"></a>}}
\if{latex}{\out{\hypertarget{method-QMatrix-new}{}}}
\subsection{Method \code{new()}}{
method to create a new object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QMatrix$new(n, names, debug = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of variables}

\item{\code{names}}{names of variables}

\item{\code{debug}}{debug flag}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QMatrix-copy"></a>}}
\if{latex}{\out{\hypertarget{method-QMatrix-copy}{}}}
\subsection{Method \code{copy()}}{
copies the values from the matrix passed as argument
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QMatrix$copy(toCopy)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{toCopy}}{matrix to be copied}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QMatrix-getProbFromTo"></a>}}
\if{latex}{\out{\hypertarget{method-QMatrix-getProbFromTo}{}}}
\subsection{Method \code{getProbFromTo()}}{
gets the probability for the presence of a link from
a variable to another
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QMatrix$getProbFromTo(from, to)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{from}}{head of link}

\item{\code{to}}{tail of link}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
probability value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QMatrix-generateParentsList"></a>}}
\if{latex}{\out{\hypertarget{method-QMatrix-generateParentsList}{}}}
\subsection{Method \code{generateParentsList()}}{
gets the list of parents analyzing the dominance
matrix formed from probabilities
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QMatrix$generateParentsList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QMatrix-getAdjacencyMatrix"></a>}}
\if{latex}{\out{\hypertarget{method-QMatrix-getAdjacencyMatrix}{}}}
\subsection{Method \code{getAdjacencyMatrix()}}{
composes the adjacency matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QMatrix$getAdjacencyMatrix()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QMatrix-getDebug"></a>}}
\if{latex}{\out{\hypertarget{method-QMatrix-getDebug}{}}}
\subsection{Method \code{getDebug()}}{
gets debug flag
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QMatrix$getDebug()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QMatrix-getDimension"></a>}}
\if{latex}{\out{\hypertarget{method-QMatrix-getDimension}{}}}
\subsection{Method \code{getDimension()}}{
gets the dimension of the matrices
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QMatrix$getDimension()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QMatrix-getIndependence"></a>}}
\if{latex}{\out{\hypertarget{method-QMatrix-getIndependence}{}}}
\subsection{Method \code{getIndependence()}}{
gets the probability for the absence of a link from
a variable to another
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QMatrix$getIndependence(first, second)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{first}}{variable of pair}

\item{\code{second}}{second variable of pair}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
probability value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QMatrix-getState"></a>}}
\if{latex}{\out{\hypertarget{method-QMatrix-getState}{}}}
\subsection{Method \code{getState()}}{
gets the state corresponding to a given position and
a certain random number
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QMatrix$getState(i, j, random)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{row of the position of interest}

\item{\code{j}}{column of the position of interest}

\item{\code{random}}{random value to get the state corresponding
to this random number}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
selected state (-1, 0, 1)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QMatrix-getValuesFromNames"></a>}}
\if{latex}{\out{\hypertarget{method-QMatrix-getValuesFromNames}{}}}
\subsection{Method \code{getValuesFromNames()}}{
gets the values for a given a position given by
variable names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QMatrix$getValuesFromNames(source, dest)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{source}}{name of the first variable}

\item{\code{dest}}{name of the second variable}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
corresponding triplet
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QMatrix-getValuesFromIndices"></a>}}
\if{latex}{\out{\hypertarget{method-QMatrix-getValuesFromIndices}{}}}
\subsection{Method \code{getValuesFromIndices()}}{
gets the values for a given a position
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QMatrix$getValuesFromIndices(i, j)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{row}

\item{\code{j}}{column}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
vector of negative, zero and positive values
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QMatrix-getVariables"></a>}}
\if{latex}{\out{\hypertarget{method-QMatrix-getVariables}{}}}
\subsection{Method \code{getVariables()}}{
gets the variables of the matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QMatrix$getVariables()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QMatrix-modelInitialization"></a>}}
\if{latex}{\out{\hypertarget{method-QMatrix-modelInitialization}{}}}
\subsection{Method \code{modelInitialization()}}{
initializes matrices with an adjacencies matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QMatrix$modelInitialization(matrix)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{matrix}}{an adjacency matrix obtained from bnlearn}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QMatrix-normalizeWithScores"></a>}}
\if{latex}{\out{\hypertarget{method-QMatrix-normalizeWithScores}{}}}
\subsection{Method \code{normalizeWithScores()}}{
method for getting a new set of probability values
compute from a score matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QMatrix$normalizeWithScores(scoreMatrix)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scoreMatrix}}{matrix used for setting new
probability values}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QMatrix-randomInitialization"></a>}}
\if{latex}{\out{\hypertarget{method-QMatrix-randomInitialization}{}}}
\subsection{Method \code{randomInitialization()}}{
method for producing a random initialization
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QMatrix$randomInitialization()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QMatrix-save"></a>}}
\if{latex}{\out{\hypertarget{method-QMatrix-save}{}}}
\subsection{Method \code{save()}}{
function for saving the data of the matrix in RDS format
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QMatrix$save(filename)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{name of the file to use}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QMatrix-setValue"></a>}}
\if{latex}{\out{\hypertarget{method-QMatrix-setValue}{}}}
\subsection{Method \code{setValue()}}{
sets the value corresponding to a given position and
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QMatrix$setValue(i, j, values, normalize = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{index of the row}

\item{\code{j}}{index of the col}

\item{\code{values}}{vector of values (negative, zero, positive)}

\item{\code{normalize}}{boolean flag (TRUE: the values must be
normalized before storing them)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QMatrix-show"></a>}}
\if{latex}{\out{\hypertarget{method-QMatrix-show}{}}}
\subsection{Method \code{show()}}{
method for producing a string with the content of the
matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QMatrix$show()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QMatrix-uniformInitialization"></a>}}
\if{latex}{\out{\hypertarget{method-QMatrix-uniformInitialization}{}}}
\subsection{Method \code{uniformInitialization()}}{
method producing a initialization with uniform values
dimension number of rows and cols
debug debug flag
variables names of variables
matrix of triplets
gets the matrix with -1, 0, 1 according to the
probability values
creates a list of triplets with the
required dimension for the problem
creates a list of values with the
required dimension for the problem
gets the max value for all the triplets
param triplets - matrix of triplets to consider
gets the sum for all the triplets
param triplets - matrix of triplets to consider
param triplets - triplet matrix of triplets
param values - list of values to substract
return - the matrix of new triplets
param - triplet matrix of triplets
param - values list of values to bivide by
return - the matrix of new triplets
apply exponential to each triplet of the matrix
passed as argument
return - the matrix of new triplets
computes the index for a given pair
of variable names
return - required index
computes the index of the triplet related
to i row and j column into matrix
return - required index
method for normalizing the values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QMatrix$uniformInitialization()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QMatrix-clone"></a>}}
\if{latex}{\out{\hypertarget{method-QMatrix-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QMatrix$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
