% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileUtils.R
\name{readCSVDataBase}
\alias{readCSVDataBase}
\title{function for reading a database}
\usage{
readCSVDataBase(filedb)
}
\arguments{
\item{filedb}{file with the data to load}
}
\value{
data frame
}
\description{
function for reading a database
}
