% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileUtils.R
\name{readDataBase}
\alias{readDataBase}
\title{function for reading a database}
\usage{
readDataBase(dbname)
}
\arguments{
\item{dbname}{file with the data to load}
}
\value{
data frame
}
\description{
function for reading a database
}
