% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphUtils.R
\name{readPositions}
\alias{readPositions}
\title{read the positions of the nodes of a graph}
\usage{
readPositions(name, path = "./data/bnet/")
}
\arguments{
\item{name}{name of the file with positions}

\item{path}{where file with positions is stored}
}
\description{
read the positions of the nodes of a graph
}
