% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreComputation.R
\name{ScoreComputation}
\alias{ScoreComputation}
\title{class for computing scores for a given data file}
\description{
class for computing scores for a given data file

class for computing scores for a given data file
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ScoreComputation-new}{\code{ScoreComputation$new()}}
\item \href{#method-ScoreComputation-getVariableName}{\code{ScoreComputation$getVariableName()}}
\item \href{#method-ScoreComputation-getVariableIndex}{\code{ScoreComputation$getVariableIndex()}}
\item \href{#method-ScoreComputation-getData}{\code{ScoreComputation$getData()}}
\item \href{#method-ScoreComputation-getNumberVariables}{\code{ScoreComputation$getNumberVariables()}}
\item \href{#method-ScoreComputation-getVariableNames}{\code{ScoreComputation$getVariableNames()}}
\item \href{#method-ScoreComputation-setDebugInfoObject}{\code{ScoreComputation$setDebugInfoObject()}}
\item \href{#method-ScoreComputation-bnlearnScore}{\code{ScoreComputation$bnlearnScore()}}
\item \href{#method-ScoreComputation-bnlearnScoreWithParents}{\code{ScoreComputation$bnlearnScoreWithParents()}}
\item \href{#method-ScoreComputation-score}{\code{ScoreComputation$score()}}
\item \href{#method-ScoreComputation-scoreWithParentSet}{\code{ScoreComputation$scoreWithParentSet()}}
\item \href{#method-ScoreComputation-show}{\code{ScoreComputation$show()}}
\item \href{#method-ScoreComputation-clone}{\code{ScoreComputation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScoreComputation-new"></a>}}
\if{latex}{\out{\hypertarget{method-ScoreComputation-new}{}}}
\subsection{Method \code{new()}}{
class constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreComputation$new(filename)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{name of the file containing the data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScoreComputation-getVariableName"></a>}}
\if{latex}{\out{\hypertarget{method-ScoreComputation-getVariableName}{}}}
\subsection{Method \code{getVariableName()}}{
gets the name of the variable with a given index
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreComputation$getVariableName(index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{index of interest}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
name of the variable
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScoreComputation-getVariableIndex"></a>}}
\if{latex}{\out{\hypertarget{method-ScoreComputation-getVariableIndex}{}}}
\subsection{Method \code{getVariableIndex()}}{
gets the index of a variable given its name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreComputation$getVariableIndex(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{of interest}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
index of the variable
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScoreComputation-getData"></a>}}
\if{latex}{\out{\hypertarget{method-ScoreComputation-getData}{}}}
\subsection{Method \code{getData()}}{
return the data used for learning
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreComputation$getData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScoreComputation-getNumberVariables"></a>}}
\if{latex}{\out{\hypertarget{method-ScoreComputation-getNumberVariables}{}}}
\subsection{Method \code{getNumberVariables()}}{
gets the number of variables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreComputation$getNumberVariables()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScoreComputation-getVariableNames"></a>}}
\if{latex}{\out{\hypertarget{method-ScoreComputation-getVariableNames}{}}}
\subsection{Method \code{getVariableNames()}}{
gets the names of variables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreComputation$getVariableNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScoreComputation-setDebugInfoObject"></a>}}
\if{latex}{\out{\hypertarget{method-ScoreComputation-setDebugInfoObject}{}}}
\subsection{Method \code{setDebugInfoObject()}}{
sets the object used for storing debug info
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreComputation$setDebugInfoObject(debugInfo)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{debugInfo}}{object to use for storing debug information}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScoreComputation-bnlearnScore"></a>}}
\if{latex}{\out{\hypertarget{method-ScoreComputation-bnlearnScore}{}}}
\subsection{Method \code{bnlearnScore()}}{
computes bnlearn score for comparison reasons
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreComputation$bnlearnScore(indexV, indexesP)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{indexV}}{index of target variable}

\item{\code{indexesP}}{indexes of parent variables}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScoreComputation-bnlearnScoreWithParents"></a>}}
\if{latex}{\out{\hypertarget{method-ScoreComputation-bnlearnScoreWithParents}{}}}
\subsection{Method \code{bnlearnScoreWithParents()}}{
computes bnlearn score for comparison reasons
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreComputation$bnlearnScoreWithParents(indexV, parents)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{indexV}}{index of target variable}

\item{\code{parents}}{vector of parent variables}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScoreComputation-score"></a>}}
\if{latex}{\out{\hypertarget{method-ScoreComputation-score}{}}}
\subsection{Method \code{score()}}{
method for computing the score of a variable given a
certain set of parents
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreComputation$score(indexV, indexesP)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{indexV}}{index of the variable of interest}

\item{\code{indexesP}}{indexes of the variables in the set of parents}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
computed score
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScoreComputation-scoreWithParentSet"></a>}}
\if{latex}{\out{\hypertarget{method-ScoreComputation-scoreWithParentSet}{}}}
\subsection{Method \code{scoreWithParentSet()}}{
method for computing the score of a variable given a
certain set of parents
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreComputation$scoreWithParentSet(target, parents)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target}}{target variable}

\item{\code{parents}}{vector of parent variables}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
computed score
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScoreComputation-show"></a>}}
\if{latex}{\out{\hypertarget{method-ScoreComputation-show}{}}}
\subsection{Method \code{show()}}{
shows information about the object
file name of the csv data file to analyze
data frame with the data in filename
variables
number of states for the variables
weights of the variables.
TODO: remove it?
scores hash for storing scores
object of class debugInfo
method for computing the number of configurations of a certain
set of variables
gets the part of the data frame for a given set
of variables (variable + parents)
computes the score for a particular subset of data
computes the weights for each variable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreComputation$show()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScoreComputation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ScoreComputation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreComputation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
