% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreHash.R
\name{ScoreHash}
\alias{ScoreHash}
\title{class definition. Data members
debug: flag for printing debug information
info: info about edges. This is needed to store
the required decoration for both sides of every edge
graph: graph with the structure of the model. Used for
making easier the computation of adjacents}
\description{
class definition. Data members
debug: flag for printing debug information
info: info about edges. This is needed to store
the required decoration for both sides of every edge
graph: graph with the structure of the model. Used for
making easier the computation of adjacents

class definition. Data members
debug: flag for printing debug information
info: info about edges. This is needed to store
the required decoration for both sides of every edge
graph: graph with the structure of the model. Used for
making easier the computation of adjacents
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ScoreTestHash-new}{\code{ScoreHash$new()}}
\item \href{#method-ScoreTestHash-store}{\code{ScoreHash$store()}}
\item \href{#method-ScoreTestHash-check}{\code{ScoreHash$check()}}
\item \href{#method-ScoreTestHash-get}{\code{ScoreHash$get()}}
\item \href{#method-ScoreTestHash-show}{\code{ScoreHash$show()}}
\item \href{#method-ScoreTestHash-setDebugInfoObject}{\code{ScoreHash$setDebugInfoObject()}}
\item \href{#method-ScoreTestHash-clone}{\code{ScoreHash$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScoreTestHash-new"></a>}}
\if{latex}{\out{\hypertarget{method-ScoreTestHash-new}{}}}
\subsection{Method \code{new()}}{
class constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreHash$new(debug = FALSE, debugInfo)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{debug}}{debug flag}

\item{\code{debugInfo}}{object for storing debug info}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScoreTestHash-store"></a>}}
\if{latex}{\out{\hypertarget{method-ScoreTestHash-store}{}}}
\subsection{Method \code{store()}}{
method to store tests information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreHash$store(variable, parents, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{variable of interest}

\item{\code{parents}}{set of parents}

\item{\code{value}}{value of the score}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScoreTestHash-check"></a>}}
\if{latex}{\out{\hypertarget{method-ScoreTestHash-check}{}}}
\subsection{Method \code{check()}}{
method to check the existance of a score information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreHash$check(variable, parents)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{variable of interest}

\item{\code{parents}}{set of parents}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
boolean value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScoreTestHash-get"></a>}}
\if{latex}{\out{\hypertarget{method-ScoreTestHash-get}{}}}
\subsection{Method \code{get()}}{
method to get the result of a score
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreHash$get(variable, parents)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{variable of interest}

\item{\code{parents}}{set of parents}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
score value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScoreTestHash-show"></a>}}
\if{latex}{\out{\hypertarget{method-ScoreTestHash-show}{}}}
\subsection{Method \code{show()}}{
show method for printing information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreHash$show()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScoreTestHash-setDebugInfoObject"></a>}}
\if{latex}{\out{\hypertarget{method-ScoreTestHash-setDebugInfoObject}{}}}
\subsection{Method \code{setDebugInfoObject()}}{
sets the object used for storing debug information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreHash$setDebugInfoObject(debugInfo)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{debugInfo}}{object for storing debug information
class private part
hash of scores
counter of calls to check
counter of calls to get
boolean flag
object of class VSLearningInfo
composes a key from the variables involved in the
score}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScoreTestHash-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ScoreTestHash-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreHash$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
